requirements.txt
absl-py==2.0.0
aiosignal==1.3.1
astunparse==1.6.3
attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
chex==0.1.85
click==8.1.7
cloudpickle==3.0.0
contextlib2==21.6.0
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
distrax==0.1.5
dm-env==1.6
dm-haiku==0.0.11
dm-tree==0.1.8
etils==1.5.2
Farama-Notifications==0.0.4
filelock==3.13.1
flatbuffers==23.5.26
flax==0.7.5
fonttools==4.47.0
frozenlist==1.4.1
fsspec==2023.12.2
gast==0.5.4
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h5py==3.10.0
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflection==0.5.1
jax==0.4.23
jaxlib==0.4.23
Jinja2==3.1.2
jmp==0.0.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
lxml==4.9.4
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
mdurl==0.1.2
ml-dtypes==0.3.1
ml_collections==0.1.1
mpmath==1.3.0
msgpack==1.0.7
nest-asyncio==1.5.8
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
open-spiel==1.4
opencv-python==4.8.1.78
opt-einsum==3.3.0
optax==0.1.7
orbax-checkpoint==0.4.8
packaging==23.2
pandas==2.1.4
pi==0.1.2
Pillow==10.1.0
protobuf==4.23.4
pyarrow==14.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pygambit==16.0.2
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
ray==2.9.0
raylib==5.0.0.0
referencing==0.32.0
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.0
rlax==0.1.6
rpds-py==0.16.2
rsa==4.9
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.35.0
tensorflow-probability==0.23.0
tensorstore==0.1.51
termcolor==2.4.0
toolz==0.12.0
torch==2.1.2
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.1.0
Werkzeug==3.0.1
wrapt==1.14.1
zipp==3.17.0
